feat(web): thin RoomPage shell with session wiring (#171)
Extract room chrome into focused hooks and presentational components so
RoomPage is layout and wiring only (460 lines). Add integration test
asserting ChatSession, SfuMediaSession, and TheaterPlayback teardown on
unmount.
